Retrieve a project and create a task.
Modified the task model to json to display the project Id of the project that it belongs to.
Changed the not found for a try except.
Now you can retrieve a project with all of its tasks.
Now you can create a task and have it automatically be part of the project of the url.